Update mast_controller_and_monitor.cpp
Implemented flags, but the system crashes in certain cases. One of these can be reliably replicated by:
1. Sliding axis_1 off-center and causing the other two to lock, then:
2. Sliding axis_2 and axis_fingers off-center, then:
3. Sliding axis_1 back to neutral.
The other two axes are thus locked and can't be placed back to neutral. 
Idea for fix: instead of checking the flag on READING the input from the axes, check for the flag on the OUTPUT stage. To be tested.